sdk/ai/azure-ai-projects/dev_requirements.txt
-e ../../../eng/tools/azure-sdk-tools
../../core/azure-core
../../identity/azure-identity
aiohttp
python-dotenv
azure.ai.inference
openai
opentelemetry-sdk
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry

